+++ master/api/README.md
@@ -1,0 +1,3 @@
+#### 初始化
++ 执行`npm install`补全需要的包
++ 在浏览器地址栏输入`http://localhost:3030/接口地址`访问接口
