+++ master/sql/curl/package.json
@@ -4,6 +4,7 @@
   "description": "",
   "main": "app.js",
   "dependencies": {
+    "express": "^4.17.1",
     "mysql": "^2.18.1"
   },
   "devDependencies": {},
